requirements.txt
boto3
psycopg2-binary
python-dotenv
redis
fuzzywuzzy
python-Levenshtein
civis